fixed _clear_flag() throwing exception when flag not exists

+++ src/scalarizr/handlers/lifecycle.py
@@ -435,7 +435,8 @@ class LifeCycleHandler(scalarizr.handlers.Handler):
 		return os.path.exists(self._get_flag_filename(name))
 	
 	def _clear_flag(self, name):
-		os.remove(self._get_flag_filename(name))	
+		if self._flag_exists(name):
+			os.remove(self._get_flag_filename(name))	
 
 	def _get_flag_contents(self, name):
 		file = self._get_flag_filename(name)
